feat(frontend): implement readonly organization structure with optimized layout and spacing
